Add clearance around bushings, default to 15mm length
When printing with small rail_d, it became clear that between
printing imprecision and bearing size imprecision, it was possible
for the rail to bind due to incompatible Z constraint.  Add a
channel in the base to accomodate this imprecision and avoid
excess constraint.  Remove material from the bottom of the clip
to expose the channel and make it effective.

Reduce the default rail_d as the previous diameter was too large
and allowed too much slop.  This was resulting in X axis yaw during
Y axis reversal.

Note that the new bushing_clearance parameter is intentionally
not accounted for in the clip sizing, since that is intended to
flex and grip the bushing firmly.