Add allowance for variance to clip profile
If the clips are too tight (due to imprecision in either the print
or the bushings) they will push the bushings up away from the base
and out of alignment with the rail holes, causing the bushings to
bind.  Apply the bushing clearance to the clips as well as the base.

Also, if the bushing is too tight end to end, the end blocks will
flex and push the bushing out, with the same end result.  A little
extra length will not cause trouble, so add it.